移除自動練習頁面；新增自訂練習元件以處理題目作答邏輯；優化題目長度計算函式；更新進行中訓練頁面以使用自訂練習元件
