+++ src/components/training/item/FillInTheBlankItem.jsx
