+++ src/components/training/item/MultipleChoiceItem.jsx
