+++ src/components/training/item/VocabularyItem.jsx
