--- src/pages/TrainingAuto.jsx
@@ -1,9 +1,0 @@
-// import { useContext } from 'react'
-// import { DataContext } from '@/context/DataContext'
-
-export default function TrainingAuto() {
-//   const { problems, setQuestions } = useContext(DataContext)
-
-
-  return <></>
-}
